refactor: move and group modules hooks and sdk
