+++ packages/functions/src/hooks/schemas/satellite.env.ts
