+++ packages/functions/src/index.ts
@@ -1,13 +1,13 @@
 export * from './hooks/db/assertions';
 export * from './hooks/db/hooks';
+export * from './hooks/schemas/collections';
+export * from './hooks/schemas/context';
+export * from './hooks/schemas/db/context';
+export * from './hooks/schemas/db/payload';
+export * from './hooks/schemas/satellite.env';
 export * from './schemas/core';
-export * from './schemas/db/context';
-export * from './schemas/db/sdk';
-export * from './schemas/db/store';
-export * from './schemas/hooks/collections';
-export * from './schemas/hooks/context';
-export * from './schemas/hooks/satellite.env';
-export * from './sdk/datastore.sdk';
+export * from './sdk/db.sdk';
+export * from './sdk/schemas/db';
 export * from './sdk/serializer.sdk';
 
 import './polyfills/console.polyfill';
